// next-ui/src/generated/openapi/types.gen.ts
export type MetadataSearchResult = {
    author?: string;
    coverUrl?: string;
    description?: string;
    externalId: string;
    provider: string;
    status?: string;
    tags: Array<string>;
    title: string;
    year?: number;
};
// ...
export type PluginApplyAuthor = {
    name: string;
    role: string;
};

export type PluginApplyCoverRequest = {
    coverUrl: string;
};

export type PluginApplyMetadataRequest = {
    ageRating?: number;
    alternativeTitles?: {
        [key: string]: string;
    };
    authors?: Array<PluginApplyAuthor>;
    externalId?: string;
    genres?: Array<string>;
    provider?: string;
    publisher?: string;
    releaseDate?: string;
    status?: string;
    summary?: string;
    tags?: Array<string>;
    title?: string;
};